b2b_sdp_demux: Cast time_t to int
CID #417533

+++ modules/b2b_sdp_demux/b2b_sdp_demux.c
@@ -1876,7 +1876,7 @@ static void b2b_sdp_server_event_trigger_create(struct b2b_sdp_ctx *ctx, bin_pac
 
 	bin_push_str(store, &ctx->callid);
 	bin_push_int(store, ctx->clients_no);
-	bin_push_int(store, ctx->sess_id);
+	bin_push_int(store, (int)(unsigned long)ctx->sess_id);
 	bin_push_str(store, &ctx->sess_ip);
 
 	list_for_each(c, &ctx->clients) {
